Add htmlproofer to test the site integrity
In order to test the site actually works correctly, and that the
content is well formed, opposed to that it just builds in Jekyll, we
can run `htmlproofer`, which will test images, links, and other details.

This will give a better early warning system, and will make it possible
to fix issues on a PR (via Travis) opposed to at a later point.