Only check against 4xx errors
If a site we're linking to has some downtime, i.e. for a database
connection issue, then that shouldn't stop Hacksoc's site from
deploying. We only want to block our site from being marked as build if
we're i.e. getting a NOT_FOUND or FORBIDDEN request because we've done
something wrong.